Register stateless core tools at ash activate, not AgentLoop start
bash/read_file/write_file/edit_file/grep/glob/ls/list_skills/pwsh used
to register inside AgentLoop's constructor, which only runs when the
backend's start() callback fires — after all extensions have loaded.
Extensions that look up tools at activate time (e.g. a scheme bridge
binding bash into its runtime) saw an empty registry and failed with
"tool 'bash' not registered".

Move those registrations into agentBackend's activate via
ctx.agent.registerTool, before extension load. Only conversation_recall
stays in AgentLoop — it needs the per-session conversation state.

The shared FileReadCache moves alongside the read_file tool and is
exposed via the `agent:file-read-cache` handler so AgentLoop can still
invalidate entries on compaction and write-tool side effects.
